add jobs functionalities
