+++ app/Http/Controllers/Api/V1/JobController.php
@@ -171,41 +171,5 @@ class JobController extends Controller
             'message' => 'Job listing deleted successfully.'
         ], Response::HTTP_OK);
     }
-    public function search(Request $request)
-{
-    $validator = Validator::make($request->all(), [
-        'query' => 'required|string|min:3',
-        'page' => 'nullable|integer|min:1',
-        'size' => 'nullable|integer|min:1',
-    ]);
-
-    if ($validator->fails()) {
-        return response()->json([
-            'success' => false,
-            'message' => 'Invalid input parameters.',
-            'errors' => $validator->errors(),
-        ], Response::HTTP_BAD_REQUEST);
-    }
-
-    $query = $request->input('query');
-    $perPage = $request->input('size', 15);
-
-    $jobs = Job::where('job_title', 'LIKE', "%{$query}%")
-                ->orWhere('description', 'LIKE', "%{$query}%")
-                ->orWhere('location', 'LIKE', "%{$query}%")
-                ->paginate($perPage);
-
-    return response()->json([
-        'message' => 'Job search results retrieved successfully.',
-        'data' => $jobs->items(),
-        'pagination' => [
-            'current_page' => $jobs->currentPage(),
-            'total_pages' => $jobs->lastPage(),
-            'page_size' => $jobs->perPage(),
-            'total_items' => $jobs->total(),
-        ],
-    ], Response::HTTP_OK);
-}
-
 
 }
